feat(adk): add eino adk examples (#114)

+++ flow/agent/multiagent/plan_execute/tools/data.go
@@ -16,7 +16,7 @@
 
 package tools
 
-var locationAdjacencyMap = map[string]map[string]int{
+var locationAdjacencyMap = map[string]map[string]float64{
 	"奇幻园林": {
 		"幻想世界": 10,
 		"未来世界": 7,
